pyproject.toml
[project]
name = "serobot"
version = "0.1.0"
description = "sero telegram bot"
authors = [
    {name = "çağatay kılınç",email = "cgtyklnc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "gmft (>=0.4.0,<0.5.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "transformers (>=4.49.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
